+++ hedera-mirror-rosetta/app/domain/types/amount_test.go
@@ -21,13 +21,27 @@
 package types
 
 import (
+	"github.com/coinbase/rosetta-sdk-go/types"
+	"github.com/hashgraph/hedera-mirror-node/hedera-mirror-rosetta/config"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
+func expectedAmount() *types.Amount {
+	return &types.Amount{
+		Value:    "400",
+		Currency: config.CurrencyHbar,
+		Metadata: nil,
+	}
+}
+
+func exampleAmount() *Amount {
+	return &Amount{Value: int64(400)}
+}
+
 func TestToRosettaAmount(t *testing.T) {
 	// when:
-	result := exampleAmount().ToRosettaAmount()
+	result := exampleAmount().ToRosetta()
 
 	// then:
 	assert.Equal(t, expectedAmount(), result)
